% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binarySet.R
\name{binarySet}
\alias{binarySet}
\title{Search for supersets and subsets}
\usage{
binarySet(profile_data)
}
\arguments{
\item{profile_data}{the binary drug-target interaction matrix with row indexes as drugs and column
indexes as targets.}
}
\value{
A list contains the following components:
\item{superset}{all the possible supersets of the input drug-target interaction data}
\item{subset}{all the possible subsets of the input drug-target interaction data}
}
\description{
A function for searching the supersets and subsets of the binary drug-target interaction data.
}
\examples{
data(tyner_interaction_binary)
sets<-binarySet(tyner_interaction_binary[1, 1:3])
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

