% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/davis.R
\name{davis}
\alias{davis}
\title{Drug-target profile for 72 drugs and 442 targets.}
\description{
Binding results (Kd's in nM) for 72 drugs vs 442 kinase assays.
Blank fields indicate interactions that were not detected in a 10 uM primary screen.
}
\references{
Davis et al. Comprehensive analysis of kinase inhibitor selectivity. Nat. Biotechnol. 2011 29, 1046-51.
}

