% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findSameCol.R
\name{findSameCol}
\alias{findSameCol}
\title{Find the same column from a matrix}
\usage{
findSameCol(X, Y)
}
\arguments{
\item{X}{a matrix}

\item{Y}{a vector with the same length as each column in X.}
}
\value{
a vector of the column indexes which are the same as vector Y.
}
\description{
A function to seek for the same column from a matrix
}
\examples{
\dontrun{
data(tyner_interaction_binary)
data(tyner_sensitivity)
x<-data.frame(tyner_interaction_binary)
kinase_names<-dimnames(tyner_interaction_binary)
float<-sffsBinary(tyner_interaction_binary, tyner_sensitivity[,1])
k_select <- float$k_sel
select_kinase_names <- findSameSet(x, k_select, kinase_names)
}
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

