% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/floating2.R
\name{floating2}
\alias{floating2}
\title{Filter targets}
\usage{
floating2(profile, sens, sp = 1, max_k = 2, verbosity = FALSE)
}
\arguments{
\item{profile}{drug-target interaction data}

\item{sens}{drug sensitivity data}

\item{sp}{an integer to specify the starting point for the sffs search algorithm. The number cannot be larger than the total number of targets in the drug-target interaction data.
By default, the starting point is the first target, namely, sp = 1.}

\item{max_k}{an integer to specify the maximal number of targets that can be selected by the sffs algorithm. In practice it is advised to keep it under 10
as the number of sensitivities to be predicted will increase exponentially. By default, max_k = 2.}

\item{verbosity}{a boolean value to decide if the information should be displayed. If it is TRUE, the information
will be displayed while the model is running. Otherwise, the information will not be displayed. By default, it is
FALSE.}
}
\value{
A list containing the following components:
\item{timma}{a list contains: the predicted efficacy matrix, prediction error and predicted drug sensitivity}
\item{k_sel}{the indexes for selected targets}
}
\description{
A function to filter targets based on their corration with the drug sensitivity
}
\details{
The major difference between original and modified averaging method is the averaging methods for the case where the minimization and maximization rules are not simultaneously satisfied.
For example, for a queried target set there are supersets but not subsets in the training data, the original algorithm will take the prediction from these supersets data using the minimization rule.
However, the modified algorithm will further adjust the prediction using the average between such a prediction and 0.
}
\examples{
\dontrun{
data(tyner_interaction_binary)
data(tyner_sensitivity)
result<-floating2(tyner_interaction_binary, tyner_sensitivity[,1], sp = 1, max_k = 5)
}
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

