% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBinary1}
\alias{getBinary1}
\title{Weighted binary set for multiclass data}
\usage{
getBinary1(input, data)
}
\arguments{
\item{input}{a vector of multiclass data}

\item{data}{a matrix of multiclass data as training data}
}
\value{
a list of the following components:
\item{superset}{the weighted supersets of the input data from the training data}
\item{subset}{the weighted subsets of the input data from the training data}
}
\description{
A function to get the weighted supersets and subsets for multiclass data
}
\examples{
data(tyner_interaction_multiclass)
sets<-getBinary1(tyner_interaction_multiclass[1,], tyner_interaction_multiclass)
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

