% Generated by roxygen2 (4.0.2): do not edit by hand
\name{graycode2}
\alias{graycode2}
\title{Graycode Function}
\usage{
graycode2(a)
}
\arguments{
\item{a}{the number of targets}
}
\value{
A list contains the following components:
\item{rows}{the number of rows}
\item{cols}{the number of columns}
\item{dec}{the decimal graycode results}
}
\description{
A function to generate decimal graycode
}
\examples{
code<-graycode2(5)
}
\author{
Liye He \email{liye.he@helsinki.fi}
}
\references{
Dah jyh Guan. (Scientific Note) Generalized Gray Codes with Applications. 1998
}

