% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graycode3.R
\name{graycode3}
\alias{graycode3}
\title{Gray code function for matrix indexes}
\usage{
graycode3(m)
}
\arguments{
\item{m}{an integer to specify the number of bits}
}
\value{
a list of the following components:
\item{gc_row}{binary gray code as row names of the predicted sensitivity matrix}
\item{gc_col}{binary gray code as column names of the predicted sensitivity matrix}
\item{dec_row}{decimal gray code as row names of the predicted sensitivity matrix}
\item{dec_col}{decimal gray code as column names of the predicted sensitivity matrix}
}
\description{
A function to generate gray code used for matrix row and column names
}
\examples{
names<-graycode3(3)
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

