% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kiba}
\alias{kiba}
\title{Kiba interaction data}
\format{A data frame with column names as target ID and the first column as CHEMBL ID.}
\description{
The curated drug-target interactions data including 52498 ChEMBL compounds and 467 targets
from the Tang et al. (2014) study
}

