% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miller_drugs.R
\name{miller_drugs}
\alias{miller_drugs}
\title{A drug list from Miller study}
\format{A data frame with drug information from the Miller study}
\description{
The drug list from the Miller study
}

