% Generated by roxygen2 (4.0.2): do not edit by hand
\name{miller_targets}
\alias{miller_targets}
\title{The curated drug-target data for the Miller drugs}
\format{A data frame contains 234 targets information for 14 drugs}
\description{
The curated drug-target data for the Miller drugs
}

