% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tyner_interaction_multiclass.R
\name{tyner_interaction_multiclass}
\alias{tyner_interaction_multiclass}
\title{A multi-class drug-target interaction data}
\format{A matrix with drug names as row names and target names as column names}
\source{
The orinial multi-class data can be found: \url{http://cancerres.aacrjournals.org/content/73/1/285/suppl/DC1}
}
\description{
A dataset containing 65 drugs and 322 targets interaction data. The binding affinity is categorical
values. A higher value indicates a stronger interaction.
}

