% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tyner_sensitivity.R
\name{tyner_sensitivity}
\alias{tyner_sensitivity}
\title{The drug sensitivity data}
\format{A matrix contains the normalized 151 patient drug sensitivity data}
\source{
The orinial 151 patient drug sensitivity data can be found: \url{http://cancerres.aacrjournals.org/content/73/1/285/suppl/DC1}
}
\description{
A dataset containing the normalized 151 patient drug sensitivity data.
}

