% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dec2bin.R
\name{dec2bin}
\alias{dec2bin}
\title{Convert decimal values to binary values}
\usage{
dec2bin(number, bits)
}
\arguments{
\item{number}{the decimal number that needs to be converted.}

\item{bits}{the number of bits of the result}
}
\value{
a vector contains of binary values 0 and 1.
}
\description{
A function to convert decimal values to binary values
}
\examples{
dec2bin(8, 5)
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

