% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getBinary.R
\name{getBinary}
\alias{getBinary}
\title{Binary set for multiclass data}
\usage{
getBinary(input, data)
}
\arguments{
\item{input}{a vector of multiclass data}

\item{data}{a matrix of multiclass data as training data}
}
\value{
a list of the following components:
\item{superset}{the supersets of the input data from the training data}
\item{subset}{the subsets of the input data from the training data}
}
\description{
A function to get the supersets and subsets for multiclass data
}
\examples{
data(tyner_interaction_multiclass)
sets<-getBinary(tyner_interaction_multiclass[1,], tyner_interaction_multiclass)
}
\author{
Liye He \email{liye.he@helsinki.fi}
}

