% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grays.R
\name{grays}
\alias{grays}
\title{Generate gray code}
\usage{
grays(n)
}
\arguments{
\item{n}{an integer to specify the number of bits.}
}
\value{
a vector of the decimal gray code.
}
\description{
A function to generate gray code
}
\examples{
code<-grays(3)
}
\author{
Liye He \email{liye.he@helsinki.fi}
}
\references{
Dah jyh Guan. (Scientific Note) Generalized Gray Codes with Applications. 1998
}

