% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miller_drug_response.R
\name{miller_drug_response}
\alias{miller_drug_response}
\title{The single drug does-response data from the Miller study}
\format{A data frame contains the drug response from Miller study}
\description{
The single drug does-response data from the Miller study.
}

