% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miller_interaction_binary.R
\name{miller_interaction_binary}
\alias{miller_interaction_binary}
\title{The binarized drug-target data for the Miller drugs}
\format{A data frame contains drug names, target names, and binding affities}
\description{
The binarized drug-target data for the Miller drugs
}

