% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miller_sensitivity.R
\name{miller_sensitivity}
\alias{miller_sensitivity}
\title{The scaled drug sensitivity data for the Miller drugs}
\format{A matrix with drugs as row indexes and entries are drug sensitivities}
\description{
The scaled drug sensitivity data for the Miller drugs
}

