% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miller_targets.R
\name{miller_targets}
\alias{miller_targets}
\title{The curated drug-target data for the Miller drugs}
\format{A data frame contains 234 targets information for 14 drugs}
\description{
The curated drug-target data for the Miller drugs
}

